test: split fixture check coverage helpers
