+++ crates/vize/src/commands/check/nuxt/fallback.rs
@@ -5,6 +5,8 @@
 //! commented-out entries and module names inside unrelated strings no longer
 //! count as installed modules.
 
+mod fallback_values;
+
 use std::path::Path;
 
 use oxc_allocator::Allocator;
@@ -21,6 +23,7 @@ use super::stubs::{
     declared_name, push_generic_function_stub, push_named_overload_stubs, push_stub,
     tracked_read_to_string,
 };
+use fallback_values::fallback_value_stubs;
 
 pub(super) fn collect_fallback_stubs(
     stubs: &mut Vec<String>,
@@ -312,101 +315,6 @@ fn fallback_type_alias_stubs() -> Vec<String> {
         "type MaybeRef<T = any> = import('vue').MaybeRef<T>;".into(),
         "type MaybeRefOrGetter<T = any> = import('vue').MaybeRefOrGetter<T>;".into(),
         "type Component = import('vue').Component;".into(),
-    ]
-}
-
-/// Hardcoded `any`-typed value stubs. Skipped whenever the project ships a
-/// generated `.nuxt` import manifest, which covers all of these names with
-/// real `typeof import(...)` types.
-fn fallback_value_stubs() -> Vec<String> {
-    vec![
-        "declare function ref<T>(value: T): Ref<UnwrapRef<T>>;".into(),
-        "declare function ref<T = any>(): Ref<T | undefined>;".into(),
-        "declare function computed<T>(getter: () => T): ComputedRef<T>;".into(),
-        "declare function computed<T>(options: { get: () => T; set: (value: T) => void }): WritableComputedRef<T>;".into(),
-        "declare function reactive<T extends object>(target: T): UnwrapNestedRefs<T>;".into(),
-        "declare function readonly<T extends object>(target: T): Readonly<T>;".into(),
-        "declare function watch(source: any, cb: (...args: any[]) => any, options?: any): any;".into(),
-        "declare function watchEffect(effect: () => void, options?: any): any;".into(),
-        "declare function watchPostEffect(effect: () => void): any;".into(),
-        "declare function watchSyncEffect(effect: () => void): any;".into(),
-        "declare function onMounted(hook: () => any): void;".into(),
-        "declare function onUnmounted(hook: () => any): void;".into(),
-        "declare function onBeforeMount(hook: () => any): void;".into(),
-        "declare function onBeforeUnmount(hook: () => any): void;".into(),
-        "declare function onBeforeUpdate(hook: () => any): void;".into(),
-        "declare function onUpdated(hook: () => any): void;".into(),
-        "declare function onActivated(hook: () => any): void;".into(),
-        "declare function onDeactivated(hook: () => any): void;".into(),
-        "declare function onErrorCaptured(hook: (...args: any[]) => any): void;".into(),
-        "declare function nextTick(fn?: () => void): Promise<void>;".into(),
-        "declare function toRef<T extends object, K extends keyof T>(object: T, key: K): Ref<T[K]>;".into(),
-        "declare function toRefs<T extends object>(object: T): { [K in keyof T]: Ref<T[K]> };".into(),
-        "declare function unref<T>(ref: T | Ref<T>): T;".into(),
-        "declare function isRef(value: any): value is Ref;".into(),
-        "declare function shallowRef<T>(value: T): ShallowRef<T>;".into(),
-        "declare function triggerRef(ref: ShallowRef): void;".into(),
-        "declare function provide<T>(key: string | symbol, value: T): void;".into(),
-        "declare function inject<T>(key: string | symbol): T | undefined;".into(),
-        "declare function inject<T>(key: string | symbol, defaultValue: T): T;".into(),
-        "declare function defineAsyncComponent(source: any): any;".into(),
-        "declare function h(type: any, ...args: any[]): any;".into(),
-        "declare function useAttrs(): Record<string, unknown>;".into(),
-        "declare function useSlots(): Record<string, (...args: any[]) => any>;".into(),
-        "declare function toRaw<T>(observed: T): T;".into(),
-        "declare function markRaw<T extends object>(value: T): T;".into(),
-        "declare function effectScope(detached?: boolean): any;".into(),
-        "declare function getCurrentScope(): any;".into(),
-        "declare function onScopeDispose(fn: () => void): void;".into(),
-        "declare function shallowReactive<T extends object>(target: T): T;".into(),
-        "declare function shallowReadonly<T extends object>(target: T): Readonly<T>;".into(),
-        "declare function customRef<T>(factory: any): Ref<T>;".into(),
-        "declare function useRouter(): any;".into(),
-        "declare function useRoute(name?: string): any;".into(),
-        "declare function definePageMeta(meta: any): void;".into(),
-        "declare function defineRouteRules(rules: any): void;".into(),
-        "declare function useSeoMeta(meta: any): void;".into(),
-        "declare function useFetch<T = any>(url: string | (() => string), options?: any): ({ data: Ref<T | null>; pending: Ref<boolean>; error: Ref<any>; refresh: () => Promise<void> } & Record<string, any>);".into(),
-        "declare function useAsyncData<T = any>(handler: (...args: any[]) => T | Promise<T>, options?: any): ({ data: Ref<T | null>; pending: Ref<boolean>; error: Ref<any>; refresh: () => Promise<void> } & Record<string, any>);".into(),
-        "declare function useAsyncData<T = any>(key: string, handler: (...args: any[]) => T | Promise<T>, options?: any): ({ data: Ref<T | null>; pending: Ref<boolean>; error: Ref<any>; refresh: () => Promise<void> } & Record<string, any>);".into(),
-        "declare function useLazyFetch<T = any>(url: string | (() => string), options?: any): ({ data: Ref<T | null>; pending: Ref<boolean>; error: Ref<any>; refresh: () => Promise<void> } & Record<string, any>);".into(),
-        "declare function useLazyAsyncData<T = any>(handler: (...args: any[]) => T | Promise<T>, options?: any): ({ data: Ref<T | null>; pending: Ref<boolean>; error: Ref<any>; refresh: () => Promise<void> } & Record<string, any>);".into(),
-        "declare function useLazyAsyncData<T = any>(key: string, handler: (...args: any[]) => T | Promise<T>, options?: any): ({ data: Ref<T | null>; pending: Ref<boolean>; error: Ref<any>; refresh: () => Promise<void> } & Record<string, any>);".into(),
-        "declare function navigateTo(to: string | any, options?: any): any;".into(),
-        "declare function createError(input: string | { statusCode?: number; statusMessage?: string; message?: string; data?: any; fatal?: boolean }): any;".into(),
-        "declare function showError(error: any): any;".into(),
-        "declare function clearError(options?: { redirect?: string }): Promise<void>;".into(),
-        "declare function useNuxtApp(): any;".into(),
-        "declare function useRuntimeConfig(): any;".into(),
-        "declare function useAppConfig(): any;".into(),
-        "declare function useState<T = any>(key: string, init?: () => T): Ref<T>;".into(),
-        "declare function useCookie<T = any>(name: string, options?: any): Ref<T>;".into(),
-        "declare function useHead(input: () => { titleTemplate?: (titleChunk?: string) => any; [key: string]: any }): void;".into(),
-        "declare function useHead(input: { titleTemplate?: (titleChunk?: string) => any; [key: string]: any }): void;".into(),
-        "declare function useHead(input: any): void;".into(),
-        "declare function useRequestHeaders(headers?: string[]): Record<string, string>;".into(),
-        "declare function useRequestURL(): URL;".into(),
-        "declare function defineNuxtComponent(options: any): any;".into(),
-        "declare function defineNuxtRouteMiddleware(middleware: any): any;".into(),
-        "declare function useError(): any;".into(),
-        "declare function abortNavigation(err?: any): any;".into(),
-        "declare function addRouteMiddleware(name: string, middleware: any, options?: any): void;".into(),
-        "declare function defineNuxtPlugin(plugin: any): any;".into(),
-        "declare function setPageLayout(layout: string): void;".into(),
-        "declare function setResponseStatus(code: number, message?: string): void;".into(),
-        "declare function prerenderRoutes(routes: string | string[]): void;".into(),
-        "declare function refreshNuxtData(keys?: string | string[]): Promise<void>;".into(),
-        "declare function clearNuxtData(keys?: string | string[]): void;".into(),
-        "declare function reloadNuxtApp(options?: any): void;".into(),
-        "declare function callOnce(key: string, fn: () => any): Promise<void>;".into(),
-        "declare function callOnce(fn: () => any): Promise<void>;".into(),
-        "declare function onNuxtReady(callback: () => any): void;".into(),
-        "declare function preloadComponents(components: string | string[]): Promise<void>;".into(),
-        "declare function prefetchComponents(components: string | string[]): Promise<void>;".into(),
-        "declare function useRequestEvent(): any;".into(),
-        "declare function useRequestFetch(): typeof globalThis.fetch;".into(),
-        "declare function useResponseHeaders(headers?: Record<string, string>): any;".into(),
-        "declare function $fetch<T = any>(...args: any[]): Promise<T>;".into(),
     ]
 }
