+++ crates/vize_canon/src/virtual_ts/generator/mod.rs
@@ -1,18 +1,12 @@
-//! Main virtual TypeScript generation entry points.
-//!
-//! Contains the public `generate_virtual_ts` and `generate_virtual_ts_with_offsets`
-//! functions that orchestrate the full virtual TypeScript generation pipeline.
-
 mod generics;
 mod imports;
 mod options_api;
+mod setup_props;
 mod spans;
 
 use vize_croquis::{BindingType, Croquis, ScopeData, ScopeKind};
 
-use self::generics::{
-    generic_injection_point, is_ident_byte, references_any_identifier, skip_ascii_ws,
-};
+use self::generics::{generic_injection_point, references_any_identifier};
 use self::imports::{
     collect_imported_names, emit_global_component_stubs, emit_reference_type_directives,
     extract_declared_name,
@@ -21,6 +15,7 @@ use self::options_api::{
     find_default_export_targets, find_options_api_props, generate_options_api_bridge,
     generate_options_api_variables,
 };
+use self::setup_props::SetupPropsPlan;
 use self::spans::{
     DEFINE_COMPONENT_HELPER, DEFINE_COMPONENT_REF, collect_template_referenced_names,
     is_local_setup_binding, merge_overlapping_spans, rewrite_export_default_for_module_scope,
@@ -32,9 +27,8 @@ use super::{
         generate_template_context, to_safe_identifier,
     },
     props::{
-        OptionsApiPropsSource, PropsTypeEmission, add_generic_defaults,
-        collect_template_prop_names, extract_generic_names, generate_props_type,
-        generate_props_variables, generate_setup_scoped_props_artifact, strip_const_modifiers,
+        OptionsApiPropsSource, add_generic_defaults, collect_template_prop_names,
+        extract_generic_names, strip_const_modifiers,
     },
     scope::{ScopeGenerationOptions, generate_scope_closures},
     types::{VirtualTsGenerationOptions, VirtualTsOptions, VirtualTsOutput, VizeMapping},
@@ -91,89 +85,6 @@ pub fn generate_virtual_ts_with_offsets(
         options,
         VirtualTsGenerationOptions::default(),
     )
-}
-
-fn is_identifier_start_byte(b: u8) -> bool {
-    b == b'_' || b == b'$' || b.is_ascii_alphabetic()
-}
-
-fn collect_typeof_root_identifiers(source: &str) -> Vec<&str> {
-    let bytes = source.as_bytes();
-    let mut idents = Vec::new();
-    let mut from = 0usize;
-
-    while let Some(rel) = source[from..].find("typeof") {
-        let at = from + rel;
-        let before_ok = at == 0 || !is_ident_byte(bytes[at - 1]);
-        let after_keyword = at + "typeof".len();
-        let after_ok = after_keyword >= bytes.len() || !is_ident_byte(bytes[after_keyword]);
-        if !before_ok || !after_ok {
-            from = after_keyword;
-            continue;
-        }
-
-        let ident_start = skip_ascii_ws(bytes, after_keyword);
-        if ident_start >= bytes.len() || !is_identifier_start_byte(bytes[ident_start]) {
-            from = after_keyword;
-            continue;
-        }
-
-        let mut ident_end = ident_start + 1;
-        while ident_end < bytes.len() && is_ident_byte(bytes[ident_end]) {
-            ident_end += 1;
-        }
-
-        let ident = &source[ident_start..ident_end];
-        if ident != "import" {
-            idents.push(ident);
-        }
-        from = ident_end;
-    }
-
-    idents
-}
-
-fn binding_is_import(summary: &Croquis, name: &str) -> bool {
-    summary.binding_spans.get(name).is_some_and(|(start, end)| {
-        summary
-            .import_statements
-            .iter()
-            .any(|imp| *start >= imp.start && *end <= imp.end)
-    })
-}
-
-fn is_setup_value_binding(summary: &Croquis, name: &str) -> bool {
-    summary.bindings.bindings.contains_key(name) && !binding_is_import(summary, name)
-}
-
-fn define_props_type_requires_setup_scope(summary: &Croquis) -> bool {
-    let Some(type_args) = summary
-        .macros
-        .define_props()
-        .and_then(|m| m.type_args.as_ref())
-    else {
-        return false;
-    };
-    let inner_type = type_args
-        .strip_prefix('<')
-        .and_then(|s| s.strip_suffix('>'))
-        .unwrap_or(type_args.as_str());
-
-    if collect_typeof_root_identifiers(inner_type)
-        .into_iter()
-        .any(|name| is_setup_value_binding(summary, name))
-    {
-        return true;
-    }
-
-    let non_hoisted_type_names: Vec<String> = summary
-        .type_exports
-        .iter()
-        .filter(|te| !te.hoisted)
-        .map(|te| te.name.as_str().into())
-        .collect();
-    !non_hoisted_type_names.is_empty()
-        && references_any_identifier(inner_type, &non_hoisted_type_names)
 }
 
 /// Generate virtual TypeScript with Vue 3 Options API binding resolution
@@ -539,7 +450,6 @@ pub(crate) fn generate_virtual_ts_with_offsets_and_checks(
     );
     ts.push('\n');
 
-    // Props type (defined at module level so it's available inside __setup).
     // For an Options API component with no `defineProps` macro, derive a real
     // `export type Props` from its runtime `props:` option so cross-file prop
     // checking is no longer a `{}` no-op. Macro-driven props (script setup) take
@@ -550,31 +460,15 @@ pub(crate) fn generate_virtual_ts_with_offsets_and_checks(
         } else {
             None
         };
-    let defer_setup_props_type = define_props_type_requires_setup_scope(summary);
-    let module_scope_declares_props = summary
-        .type_exports
-        .iter()
-        .any(|te| te.hoisted && te.name.as_str() == "Props");
-    let component_props_type_ref = if defer_setup_props_type && module_scope_declares_props {
-        "__VizeResolvedProps"
-    } else {
-        "Props"
-    };
-    let props_type_emission = if defer_setup_props_type {
-        PropsTypeEmission::DeferredToSetup
-    } else {
-        PropsTypeEmission::Module
-    };
-    profile!(
-        "canon.virtual_ts.generate_props_type",
-        generate_props_type(
+    let setup_props_plan = SetupPropsPlan::new(summary);
+    profile!("canon.virtual_ts.generate_props_type", {
+        setup_props_plan.generate_props_type(
             &mut ts,
             summary,
             generic_param,
             options_api_props.as_ref(),
-            props_type_emission,
         )
-    );
+    });
 
     // Setup scope: function that contains setup helpers and script content
     ts.push_str("// ========== Setup Scope ==========\n");
@@ -836,12 +730,7 @@ pub(crate) fn generate_virtual_ts_with_offsets_and_checks(
         });
     }
 
-    if defer_setup_props_type {
-        profile!(
-            "canon.virtual_ts.generate_setup_scoped_props_artifact",
-            generate_setup_scoped_props_artifact(&mut ts, summary)
-        );
-    }
+    setup_props_plan.emit_artifact(&mut ts, summary);
 
     // Template scope (nested inside setup)
     if template_ast.is_some() {
@@ -912,15 +801,9 @@ pub(crate) fn generate_virtual_ts_with_offsets_and_checks(
             ts.push('\n');
 
             // Props are available in template as variables
-            profile!(
-                "canon.virtual_ts.generate_props_variables",
-                generate_props_variables(
-                    &mut ts,
-                    summary,
-                    generic_param,
-                    defer_setup_props_type.then_some("__VizeSetupProps"),
-                )
-            );
+            profile!("canon.virtual_ts.generate_props_variables", {
+                setup_props_plan.generate_props_variables(&mut ts, summary, generic_param)
+            });
             if options_api {
                 profile!(
                     "canon.virtual_ts.generate_options_api_variables",
@@ -1113,9 +996,7 @@ pub(crate) fn generate_virtual_ts_with_offsets_and_checks(
     // extracted at module level while keeping each runtime expression in setup
     // scope, where script-setup bindings are defined.
     let mut setup_return_fields = Vec::new();
-    if defer_setup_props_type {
-        setup_return_fields.push("__vize_setup_props");
-    }
+    setup_props_plan.push_return_field(&mut setup_return_fields);
     if let Some(expose) = summary.macros.define_expose()
         && expose.type_args.is_none()
         && let Some(runtime_args) = expose.runtime_args.as_ref()
@@ -1142,17 +1023,7 @@ pub(crate) fn generate_virtual_ts_with_offsets_and_checks(
     ts.push_str("// Invoke setup to verify types\n");
     ts.push_str("__setup();\n\n");
 
-    if defer_setup_props_type {
-        if module_scope_declares_props {
-            ts.push_str(
-                "type __VizeResolvedProps = Awaited<ReturnType<typeof __setup>>[\"__vize_setup_props\"];\n\n",
-            );
-        } else {
-            ts.push_str(
-                "export type Props = Awaited<ReturnType<typeof __setup>>[\"__vize_setup_props\"];\n\n",
-            );
-        }
-    }
+    setup_props_plan.emit_module_export(&mut ts);
 
     // Emits type
     let emits_already_defined = summary
@@ -1280,14 +1151,7 @@ pub(crate) fn generate_virtual_ts_with_offsets_and_checks(
     // Default export
     ts.push_str("// ========== Default Export ==========\n");
     ts.push_str("type __VizeComponentInstance = {\n");
-    if has_emits_for_props {
-        append!(
-            ts,
-            "  $props: {component_props_type_ref} & __EmitProps<Emits>;\n"
-        );
-    } else {
-        append!(ts, "  $props: {component_props_type_ref};\n");
-    }
+    setup_props_plan.emit_component_props_field(&mut ts, has_emits_for_props);
     ts.push_str("  $emit: __EmitFn<Emits>;\n");
     ts.push_str("  $slots: Slots;\n");
     if has_exposed_type {
@@ -1334,7 +1198,7 @@ pub(crate) fn generate_virtual_ts_with_offsets_and_checks(
     } else {
         ""
     };
-    if let Some(generic) = generic_param.filter(|_| !defer_setup_props_type) {
+    if let Some(generic) = setup_props_plan.generic_param(generic_param) {
         let generic_decl = add_generic_defaults(generic);
         let generic_names = extract_generic_names(generic);
         append!(
