+++ crates/vize_canon/src/virtual_ts/tests.rs
@@ -1,14 +1,12 @@
-//! Tests for virtual TypeScript generation.
-//!
-//! Extracted from the `virtual_ts` module root; kept as the `tests`
-//! submodule so insta snapshot names remain `virtual_ts::tests::*`.
-
 use super::helpers::{VUE_SETUP_HELPERS, generate_template_context, get_dom_event_type};
 use super::{
     TemplateGlobal, VirtualTsCheckOptions, VirtualTsGenerationOptions, VirtualTsOptions,
     generate_virtual_ts, generate_virtual_ts_with_offsets,
     generate_virtual_ts_with_offsets_and_checks, generate_virtual_ts_with_offsets_options_api,
 };
+
+mod define_props_scope;
+mod vif_chain;
 
 fn assert_virtual_ts_snapshot(name: &str, value: &str) {
     insta::with_settings!({
@@ -1263,42 +1261,6 @@ defineProps<{ log: Log }>()
 }
 
 #[test]
-fn test_prefixed_v_else_branch_keeps_negated_branch_guard() {
-    use vize_croquis::{Analyzer, AnalyzerOptions};
-
-    let script = r#"const isOpen = true
-type Item =
-  | { kind: 'anchor'; hash: string; key: string }
-  | { kind: 'page'; to: string; key: string }
-const navItems: Item[] = []
-"#;
-    let template = r#"<div v-if="isOpen">
-  <div v-for="item in navItems" :key="item.key">
-    <span v-if="item.kind === 'page'">{{ item.to }}</span>
-    <span v-else>{{ item.hash }}</span>
-  </div>
-</div>"#;
-
-    let allocator = vize_carton::Bump::new();
-    let (root, _) = vize_armature::parse(&allocator, template);
-
-    let mut analyzer = Analyzer::with_options(AnalyzerOptions::full());
-    analyzer.analyze_script_setup(script);
-    analyzer.analyze_template(&root);
-    let summary = analyzer.finish();
-
-    let output = generate_virtual_ts(&summary, Some(script), Some(&root), 0);
-
-    assert!(
-        output
-            .code
-            .contains("} else if ((isOpen) && !(item.kind === 'page')) {"),
-        "prefixed v-else branch should retain the negated discriminant guard:\n{}",
-        output.code
-    );
-}
-
-#[test]
 fn test_scoped_slot_expressions() {
     use vize_croquis::{Analyzer, AnalyzerOptions};
 
@@ -1813,138 +1775,6 @@ defineProps<{
     assert!(
         module_scope.contains("type Option<T extends string = any> = { key: T; label: string }"),
         "hoisted type should gain the SFC generic parameter so `T` resolves at module scope:\n{}",
-        output.code
-    );
-}
-
-#[test]
-fn test_define_props_typeof_setup_binding_deferred_to_setup_scope() {
-    use vize_croquis::{Analyzer, AnalyzerOptions};
-
-    let script = r#"const someDefinition = {
-  foo: 'fooVal',
-} as const;
-
-type SomeGenericType<T extends Record<string, unknown>> = {
-  baz: string;
-  items: T;
-};
-
-const props = defineProps<SomeGenericType<typeof someDefinition>>();
-"#;
-
-    let mut analyzer = Analyzer::with_options(AnalyzerOptions::full());
-    analyzer.analyze_script_setup(script);
-    let summary = analyzer.finish();
-
-    let output =
-        generate_virtual_ts_with_offsets(&summary, Some(script), None, 0, 0, &Default::default());
-    let (module_scope, setup_and_after) = output
-        .code
-        .split_once("// ========== Setup Scope ==========")
-        .expect("setup scope marker present");
-
-    assert!(
-        !module_scope.contains("typeof someDefinition"),
-        "setup-scope value must not be referenced from module scope:\n{}",
-        output.code
-    );
-    assert!(
-        setup_and_after.contains("type __VizeSetupProps = SomeGenericType<typeof someDefinition>;"),
-        "setup-local props artifact should preserve the concrete props type:\n{}",
-        output.code
-    );
-    assert!(
-        output.code.contains(
-            "export type Props = Awaited<ReturnType<typeof __setup>>[\"__vize_setup_props\"];"
-        ),
-        "module Props should be exported through __setup return type:\n{}",
-        output.code
-    );
-}
-
-#[test]
-fn test_deferred_define_props_does_not_redeclare_hoisted_props_type() {
-    use vize_croquis::{Analyzer, AnalyzerOptions};
-
-    let script = r#"interface Props {
-  baz: string;
-}
-
-const someDefinition = {
-  foo: 'fooVal',
-} as const;
-
-defineProps<Props & { items: typeof someDefinition }>();
-"#;
-
-    let mut analyzer = Analyzer::with_options(AnalyzerOptions::full());
-    analyzer.analyze_script_setup(script);
-    let summary = analyzer.finish();
-
-    let output =
-        generate_virtual_ts_with_offsets(&summary, Some(script), None, 0, 0, &Default::default());
-
-    assert!(
-        output.code.contains("interface Props {\n  baz: string;\n}"),
-        "existing hoisted Props should remain in module scope:\n{}",
-        output.code
-    );
-    assert!(
-        !output
-            .code
-            .contains("export type Props = Awaited<ReturnType<typeof __setup>>"),
-        "deferred props must not redeclare an existing module-scope Props:\n{}",
-        output.code
-    );
-    assert!(
-        output.code.contains(
-            "type __VizeResolvedProps = Awaited<ReturnType<typeof __setup>>[\"__vize_setup_props\"];"
-        ),
-        "default export should use a private resolved props alias:\n{}",
-        output.code
-    );
-    assert!(
-        output.code.contains("  $props: __VizeResolvedProps;"),
-        "component instance should use the full deferred props type:\n{}",
-        output.code
-    );
-}
-
-#[test]
-fn test_define_props_non_hoisted_type_ref_deferred_to_setup_scope() {
-    use vize_croquis::{Analyzer, AnalyzerOptions};
-
-    let script = r#"type WidgetProps = {
-  items: typeof widgetDefinitions;
-};
-
-const widgetDefinitions = {
-  clock: {},
-} as const;
-
-defineProps<WidgetProps>();
-"#;
-
-    let mut analyzer = Analyzer::with_options(AnalyzerOptions::full());
-    analyzer.analyze_script_setup(script);
-    let summary = analyzer.finish();
-
-    let output =
-        generate_virtual_ts_with_offsets(&summary, Some(script), None, 0, 0, &Default::default());
-    let (module_scope, setup_and_after) = output
-        .code
-        .split_once("// ========== Setup Scope ==========")
-        .expect("setup scope marker present");
-
-    assert!(
-        !module_scope.contains("WidgetProps"),
-        "non-hoisted setup type must not be referenced from module scope:\n{}",
-        output.code
-    );
-    assert!(
-        setup_and_after.contains("type __VizeSetupProps = WidgetProps;"),
-        "setup-local props artifact should reference the non-hoisted type:\n{}",
         output.code
     );
 }
