+++ tests/_helpers/apps.ts
@@ -2106,22 +2106,6 @@ export const ecosystemProductsApp: AppConfig = {
   },
 };
 
-export const genericBuildApp: AppConfig = {
-  name: "generic-build",
-  cwd: path.join(PROJECTS_DIR, "generic-build"),
-  command: "",
-  args: [],
-  port: 0,
-  url: "",
-  mountSelector: "",
-  readyPattern: /./,
-  startupTimeout: 0,
-  check: {
-    cwd: path.join(PROJECTS_DIR, "generic-build"),
-    patterns: ["src/**/*.vue"],
-  },
-};
-
 export const nuxtParityApp: AppConfig = {
   name: "nuxt-parity",
   cwd: path.join(PROJECTS_DIR, "nuxt-parity"),
@@ -2135,23 +2119,6 @@ export const nuxtParityApp: AppConfig = {
   check: {
     cwd: path.join(PROJECTS_DIR, "nuxt-parity"),
     patterns: ["pages/**/*.vue", "components/**/*.vue"],
-  },
-};
-
-export const typecheckVueImportsApp: AppConfig = {
-  name: "typecheck-vue-imports",
-  cwd: path.join(PROJECTS_DIR, "typecheck-vue-imports"),
-  command: "",
-  args: [],
-  port: 0,
-  url: "",
-  mountSelector: "",
-  readyPattern: /./,
-  startupTimeout: 0,
-  check: {
-    cwd: path.join(PROJECTS_DIR, "typecheck-vue-imports"),
-    patterns: ["src/**/*.vue"],
-    tsconfig: "tsconfig.json",
   },
 };
