+++ tests/snapshots/check/generic-build.ts
@@ -1,7 +1,8 @@
 import assert from "node:assert/strict";
 import { before, describe, it } from "node:test";
 
-import { genericBuildApp, requireVizeAndCorsaBins } from "../../_helpers/apps.ts";
+import { requireVizeAndCorsaBins } from "../../_helpers/apps.ts";
+import { genericBuildApp } from "../../_helpers/fixture-apps.ts";
 import { runCrashFreeVizeCheck } from "../_helpers/realworld.ts";
 
 const app = genericBuildApp;
